code cleanup for merge
